Reorganize core types
