+++ packages/core/src/types/index.ts
@@ -2,5 +2,8 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
+export * from './enums.js'
+export * from './steps.js'
+export * from './tables.js'
 export * from './Pipeline.js'
 export * from './TableStore.js'
